fix: moved MoxyJsonFeatureCustomJsonProvider.java into a better position
