+++ rest-api/core/src/main/java/org/eclipse/kapua/app/api/core/MoxyJsonFeatureCustomJsonProvider.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.app.api.web;
+package org.eclipse.kapua.app.api.core;
 
 import org.glassfish.jersey.internal.InternalProperties;
 import org.glassfish.jersey.internal.util.PropertiesHelper;
