+++ rest-api/web/src/main/java/org/eclipse/kapua/app/api/web/RestApisApplication.java
@@ -21,6 +21,7 @@ import javax.xml.bind.JAXBException;
 import org.eclipse.kapua.app.api.core.KapuaSerializableBodyWriter;
 import org.eclipse.kapua.app.api.core.ListBodyWriter;
 import org.eclipse.kapua.app.api.core.MoxyJsonConfigContextResolver;
+import org.eclipse.kapua.app.api.core.MoxyJsonFeatureCustomJsonProvider;
 import org.eclipse.kapua.commons.rest.errors.ExceptionConfigurationProvider;
 import org.eclipse.kapua.locator.KapuaLocator;
 import org.eclipse.kapua.locator.guice.GuiceLocatorImpl;
